feat: jpa repository for quiz_attempts + added token versions in users table

+++ Quiz-Website/src/main/java/com/freeuni/proj_100/quizwebsite/model/Quiz.java
@@ -14,11 +14,11 @@ public class Quiz {
     /** Auto-incremented primary key. */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     /** ID of the user who created this quiz. */
     @Column(name = "creator_id")
-    private int creatorId;
+    private Long creatorId;
 
     /** Title of the quiz. */
     @Column(nullable = false)
@@ -50,8 +50,8 @@ public class Quiz {
 
     public Quiz() {}
 
-    public int getId() { return id; }
-    public int getCreatorId() { return creatorId; }
+    public Long getId() { return id; }
+    public Long getCreatorId() { return creatorId; }
     public String getTitle() { return title; }
     public String getDescription() { return description; }
     public boolean isRandomized() { return isRandomized; }
@@ -60,8 +60,8 @@ public class Quiz {
     public boolean isAllowPractice() { return allowPractice; }
     public LocalDateTime getCreatedAt() { return createdAt; }
 
-    public void setId(int id) { this.id = id; }
-    public void setCreatorId(int creatorId) { this.creatorId = creatorId; }
+    public void setId(Long id) { this.id = id; }
+    public void setCreatorId(Long creatorId) { this.creatorId = creatorId; }
     public void setTitle(String title) { this.title = title; }
     public void setDescription(String description) { this.description = description; }
     public void setRandomized(boolean randomized) { isRandomized = randomized; }
